Use 1.0 character spacing in training images
Adding character spacing results in fewer box-finding failures when
training, meaning more valid character samples. This improves
recognition slightly (~0.2%), and also means that the log for a build
is not hundreds of thousands of lines long. The downside could be that
spacing information isn't generated as accurately, but testing doesn't
seem to support that.